+++ aim/web/ui/src/pages/Scatters/Scatters.tsx
@@ -101,6 +101,7 @@ function Scatters(
                 <ChartPanel
                   key={props.scatterPlotData?.[0]?.data?.length}
                   ref={props.chartPanelRef}
+                  chartPanelOffsetHeight={props.chartPanelOffsetHeight}
                   panelResizing={props.panelResizing}
                   chartType={ChartTypeEnum.ScatterPlot}
                   data={props.scatterPlotData}