+++ aim/web/ui/src/types/components/HighPlot/HighPlot.d.ts
@@ -8,9 +8,19 @@ import { CurveEnum } from 'utils/d3';
 export interface IHighPlotProps {
   index: number;
   nameKey?: string;
+  brushExtents: {
+    [key: string]: {
+      [key: string]: [number, number] | [string, string];
+    };
+  };
   curveInterpolation: CurveEnum;
   isVisibleColorIndicator: boolean;
   syncHoverState: (args: ISyncHoverStateArgs) => void;
+  onAxisBrushExtentChange: (
+    key: string,
+    extent: [number, number] | [string, string] | null,
+    chartIndex: number,
+  ) => void;
   data: any;
   chartTitle?: IChartTitle;
   readOnly?: boolean;