+++ aim/web/ui/src/types/services/models/explorer/createAppModel.d.ts
@@ -136,6 +136,11 @@ export interface IHighPlotConfig {
   isVisibleColorIndicator: boolean;
   focusedState: IFocusedState;
   tooltip: IChartTooltip;
+  brushExtents: {
+    [key: string]: {
+      [key: string]: [number, number] | [string, string];
+    };
+  };
 }
 
 export interface ILineChartConfig {